refactor(validation): move the subprocess failure classifier beside its contract
classify_subprocess_failure reads the two stderr markers _run_subprocess
produces, so it belongs in subprocess_runner.py next to the code that writes
them, not in checks_common.py which only re-exports that wrapper. Putting it in
checks_common also pushed that file from 492 to 530 lines, adding a file-size
violation and spending ratchet slack for a helper that was in the wrong module.

checks_common re-exports it alongside _run_subprocess, so the checks_* modules
keep one shared subprocess import surface and no caller changes.

Evidence: taste count ratchet back to 564 <= 566, the value before this helper
was written. 292 tests pass across the six affected suites; ruff and mypy clean
on scripts/ and tests/.

Refs #5635
