+++ scripts/validation/subprocess_runner.py
@@ -19,6 +19,14 @@ from pathlib import Path
 
 from scripts.cli_exec import resolve_executable
 
+# Reason codes for the typed evidence contract (issue #5635).
+from scripts.validation.evidence import REASON_TIMEOUT, REASON_TOOL_ABSENT
+
+# The exit code :func:`_run_subprocess` returns when the child never produced
+# one of its own: a timeout or a missing executable. Any other non-zero value
+# came from the child itself.
+_SENTINEL_EXIT = -1
+
 
 def _decode_stream(data: bytes | str | None) -> str:
     """Decode a captured subprocess stream with UTF-8 replacement semantics.
@@ -76,3 +84,30 @@ def _run_subprocess(
         marker = f"Command timed out after {timeout}s"
         combined_stderr = f"{partial_stderr}\n{marker}" if partial_stderr else marker
         return -1, partial_stdout, combined_stderr
+
+
+def classify_subprocess_failure(exit_code: int, stderr: str, *, default: str) -> str:
+    """Return the evidence reason code for a failed :func:`_run_subprocess` call.
+
+    A timed-out child and a child that ran and exited non-zero are different
+    findings with different remedies, and ``_run_subprocess`` reports both as
+    exit ``-1`` versus a real code. Collapsing them into one reason costs the
+    reader the first diagnostic step, which is the defect
+    ``.claude/rules/ci-scripts.md`` MUST 14 records for the count ratchets and
+    the reason ``evidence.py`` keeps BLOCKED and UNKNOWN apart.
+
+    The two markers are :func:`_run_subprocess`'s documented contract,
+    quoted verbatim from the source above:
+
+        marker = f"Command timed out after {timeout}s"
+        return -1, "", f"Command not found: {args[0]}"
+
+    ``default`` is the caller's own reason for an ordinary non-zero exit, since
+    only the caller knows what its child was doing (issue #5635).
+    """
+    if exit_code == _SENTINEL_EXIT:
+        if "Command timed out after" in stderr:
+            return REASON_TIMEOUT
+        if "Command not found:" in stderr:
+            return REASON_TOOL_ABSENT
+    return default
